Extract formula-response detector family
